Added Time Bin Correlation evaluations

+++ include/system/system.h
@@ -340,17 +340,17 @@ class System {
         return 2.0 * op * rho * opd - opd * op * rho - rho * opd * op;
     }
 
-    inline double getTimeOf(int index, const std::vector<int>& index_vector) {
-        const int total_index = std::accumulate(index_vector.begin(), index_vector.begin() + index, 0);
+    inline double getTimeOf(size_t index, const std::vector<size_t>& index_vector) {
+        const size_t total_index = std::accumulate(index_vector.begin(), index_vector.begin() + index, 0);
         // Edge case, we are at the end of the time vector. here, we just extrapolate the time
-        const int final_index = total_index+index_vector[index];
+        const size_t final_index = total_index+index_vector[index];
         if (final_index >= parameters.grid_values.size())
             return parameters.grid_values.back() + (final_index - parameters.grid_values.size() + 1) * parameters.grid_steps.back();
         return parameters.grid_values[total_index+index_vector[index]] - parameters.grid_values[total_index];
     }
-    inline double getDeltaTimeOf(int index, const std::vector<int>& index_vector) {
-        int total_index = std::accumulate(index_vector.begin(), index_vector.end()-1, 0);
-        total_index = std::max<int>(total_index , parameters.grid_steps.size()-1);
+    inline double getDeltaTimeOf(size_t index, const std::vector<size_t>& index_vector) {
+        size_t total_index = std::accumulate(index_vector.begin(), index_vector.end()-1, 0);
+        total_index = std::max<size_t>(total_index , parameters.grid_steps.size()-1);
         return parameters.grid_steps[total_index];
     }
 };
